updated launch screen for macos
